requirements.txt
functions-framework==3.*
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
requests==2.31.0
slack-bolt==1.18.0
slack-sdk==3.21.3
urllib3==2.0.3
python-dotenv~=1.0.0
